refactor: Rename `instrument` namespace into `instrumentation`
It's to not confuse it with `instrument` function

BREAKING CHANGE:
- Rename `instrument/*`utilities into `instrumenation/*`
- Rename `serveressSdk.instrument` into `serverlessSdk.instrumentation`
