+++ node/packages/aws-lambda-sdk/.ts-types/instrumentation/aws-sdk-v2.d.ts
