+++ node/packages/aws-lambda-sdk/.ts-types/instrumentation/aws-sdk-v3-client.d.ts
