+++ node/packages/aws-lambda-sdk/.ts-types/instrumentation/express-app.d.ts
