+++ node/packages/aws-lambda-sdk/lib/instrumentation/aws-sdk/service-mapper.js
