+++ node/packages/aws-lambda-sdk/lib/instrumentation/express/instrument-layer-prototype.js
