+++ node/packages/aws-lambda-sdk/lib/instrumentation/utils/cjs-hook.js
